requirements_local.txt
# Local Deployment Requirements
# Install with: pip install -r requirements_local.txt

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Transformers for BERT
transformers>=4.30.0

# Demo interface
gradio>=4.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# Visualization (for explainability)
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0

# Utilities
scikit-learn>=1.3.0
tqdm>=4.65.0
